Port back missing type methods
