--- src/Generator/Type/ComposedType.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Fidry\Console\Generator\Type;
-
-interface ComposedType
-{
-    public function getInnerType(): InputType;
-}
